refactor metric and task state manager from executor to runtime common
