+++ runtime/executor/src/main/java/org/apache/nemo/runtime/executor/Executor.java
@@ -34,7 +34,10 @@ import org.apache.nemo.common.ir.edge.executionproperty.DecompressionProperty;
 import org.apache.nemo.common.ir.edge.executionproperty.EncoderProperty;
 import org.apache.nemo.common.ir.vertex.IRVertex;
 import org.apache.nemo.conf.JobConf;
+import org.apache.nemo.runtime.common.MetricManagerWorker;
+import org.apache.nemo.runtime.common.MetricMessageSender;
 import org.apache.nemo.runtime.common.RuntimeIdManager;
+import org.apache.nemo.runtime.common.TaskStateManager;
 import org.apache.nemo.runtime.common.comm.ControlMessage;
 import org.apache.nemo.runtime.common.message.MessageContext;
 import org.apache.nemo.runtime.common.message.MessageEnvironment;
